updated
updated readmes and added initial research. More PRs coming in shortly
