+++ .gitignore
@@ -1,0 +1,3 @@
+# virtual envirnment folder
+/venv
+demo-1.mp4
