+++ dataset/dataset.md
@@ -1,0 +1,20 @@
+# Dataset
+## Hugging face dataset
+
+Huggingface page for acquring the dataset
+
+https://huggingface.co/datasets/keremberke/garbage-object-detection
+
+#### RoboFlow dataset for Garbage Classification
+* RoboFlow is a dataset for garbage classification. It contains 10,000 images of garbage, which are divided into 10 classes. 
+* RoboFlow platform for collecting and sharing datasets for robotics and computer vision.
+
+The dataset is suitable for training and testing models for garbage classification.
+
+https://universe.roboflow.com/material-identification/garbage-classification-3/dataset/2
+
+Format for Dataset Download:
+```Yolo V5 PTorch```
+
+***more coming soon***
+
